Reorganise DOM extensions and helpers in package
