+++ src/main/kotlin/io/hemin/wien/dom/DomParsingExtensions.kt
@@ -1,10 +1,12 @@
-package io.hemin.wien.builder
+package io.hemin.wien.dom
 
+import io.hemin.wien.builder.HrefOnlyImageBuilder
+import io.hemin.wien.builder.ITunesCategoryBuilder
+import io.hemin.wien.builder.PersonBuilder
+import io.hemin.wien.builder.RssCategoryBuilder
+import io.hemin.wien.builder.RssImageBuilder
 import io.hemin.wien.parser.DateParser
 import io.hemin.wien.util.FeedNamespace
-import io.hemin.wien.util.asListOfNodes
-import io.hemin.wien.util.findElementByName
-import io.hemin.wien.util.getAttributeValueByName
 import io.hemin.wien.util.trimmedOrNullIfBlank
 import org.w3c.dom.Element
 import org.w3c.dom.Node
