+++ src/main/kotlin/io/hemin/wien/dom/DomWritingExtensions.kt
@@ -1,64 +1,16 @@
-package io.hemin.wien.util
+package io.hemin.wien.dom
 
 import io.hemin.wien.model.Category
 import io.hemin.wien.model.HrefOnlyImage
 import io.hemin.wien.model.Person
 import io.hemin.wien.model.RssImage
+import io.hemin.wien.util.BooleanStringStyle
+import io.hemin.wien.util.FeedNamespace
+import io.hemin.wien.util.asBooleanString
 import org.w3c.dom.Attr
 import org.w3c.dom.Document
 import org.w3c.dom.Element
-import org.w3c.dom.NamedNodeMap
 import org.w3c.dom.Node
-import org.w3c.dom.NodeList
-
-/** Finds the first element matching the given (local)[name], [namespace] and [filter], if any. */
-internal fun Node.findElementByName(
-    name: String,
-    namespace: FeedNamespace? = null,
-    filter: (Node) -> Boolean = { true }
-) = childNodes.asListOfNodes()
-    .firstOrNull { it.getTagName() == name && it.namespaceURI == namespace?.uri && filter(it) }
-    ?.asElement()
-
-private fun Node.getTagName(): String? = when (this) {
-    is Element -> localName ?: tagName ?: nodeName
-    else -> localName ?: nodeName
-}
-
-/**
- * Checks whether the node is a direct child of a tag with the given name.
- */
-internal fun Node.isDirectChildOf(tagName: String) =
-    parentNode.nodeName == tagName && parentNode.namespaceURI == null
-
-/**
- * Extract the [`textContent`][Node.getTextContent] of a DOM node attribute identified by name.
- *
- * @param attributeName The name of the node's attribute.
- * @param namespace The namespace to use, if any.
- * @return The textContent of the node's attribute.
- */
-internal fun Node.getAttributeValueByName(attributeName: String, namespace: FeedNamespace? = null): String? =
-    attributes?.getNamedItemNS(namespace?.uri, attributeName)?.textContent?.trim()
-
-/** Returns true if the [NodeList] contains at least one node. */
-internal fun NodeList.isNotEmpty() = length > 0
-
-/** Converts this [NodeList] to a [List] of [Node]s. */
-internal fun NodeList.asListOfNodes(): List<Node> {
-    if (length == 0) return emptyList()
-    return NodeListWrapper(this)
-}
-
-/** Converts this [NamedNodeMap] to a [List] of [Attr]s. */
-internal fun NamedNodeMap.asListOfAttrs(): List<Attr> {
-    if (length == 0) return emptyList()
-    return (0 until length).map { index -> item(index) }
-        .filterIsInstance(Attr::class.java)
-}
-
-/** Casts the [Node] to an [Element]. */
-internal fun Node.asElement() = this as Element
 
 /**
  * Appends a new `<namespace:image href="..."/>` with the given [namespace] to this [Element].
@@ -93,7 +45,7 @@ internal fun Node.appendRssImageElement(image: RssImage): Element = appendElemen
 }
 
 /**
- * Appends a new [Element] with the given [tag name][elementTagName] to this [Element].
+ * Appends a new [Element] with the given [tag name][elementTagName] to this [Node].
  * You can configure the new element by specifiying a lambda for [init].
  *
  * @param elementTagName The tag name for the new element.
@@ -168,7 +120,7 @@ internal fun Node.appendTrueFalseElement(tagName: String, value: Boolean, namesp
  * @param person The person instance to use
  * @param namespace The namespace to use, if any
  */
-internal fun Element.appendPersonElement(tagName: String, person: Person, namespace: FeedNamespace? = null) {
+internal fun Node.appendPersonElement(tagName: String, person: Person, namespace: FeedNamespace? = null) {
     appendElement(tagName, namespace) {
         appendElement("name", namespace) { textContent = person.name }
 
@@ -188,7 +140,7 @@ internal fun Element.appendPersonElement(tagName: String, person: Person, namesp
  * @param categories The [categories][Category.ITunes] to append.
  * @param namespace The namespace to use, if any.
  */
-internal fun Element.appendITunesCategoryElements(categories: List<Category.ITunes>, namespace: FeedNamespace? = null) {
+internal fun Node.appendITunesCategoryElements(categories: List<Category.ITunes>, namespace: FeedNamespace? = null) {
     for (category in categories) {
         appendElement("category", namespace) {
             setAttribute("text", category.category)
@@ -208,7 +160,7 @@ internal fun Element.appendITunesCategoryElements(categories: List<Category.ITun
  * @param categories The [categories][Category.Rss] to append.
  * @param namespace The namespace to use, if any.
  */
-internal fun Element.appendRssCategoryElements(categories: List<Category.Rss>, namespace: FeedNamespace? = null) {
+internal fun Node.appendRssCategoryElements(categories: List<Category.Rss>, namespace: FeedNamespace? = null) {
     for (category in categories) {
         appendElement("category", namespace) {
             textContent = category.category
