+++ src/main/kotlin/io/hemin/wien/parser/namespace/FeedpressParser.kt
@@ -2,7 +2,7 @@ package io.hemin.wien.parser.namespace
 
 import io.hemin.wien.builder.episode.EpisodeBuilder
 import io.hemin.wien.builder.podcast.PodcastBuilder
-import io.hemin.wien.builder.textOrNull
+import io.hemin.wien.dom.textOrNull
 import io.hemin.wien.parser.NamespaceParser
 import io.hemin.wien.util.FeedNamespace
 import org.w3c.dom.Node
