+++ src/main/kotlin/io/hemin/wien/parser/namespace/GooglePlayParser.kt
@@ -2,13 +2,12 @@ package io.hemin.wien.parser.namespace
 
 import io.hemin.wien.builder.episode.EpisodeBuilder
 import io.hemin.wien.builder.podcast.PodcastBuilder
-import io.hemin.wien.builder.textAsBooleanOrNull
-import io.hemin.wien.builder.textOrNull
-import io.hemin.wien.builder.toHrefOnlyImageBuilder
-import io.hemin.wien.builder.toITunesCategoryBuilder
+import io.hemin.wien.dom.textAsBooleanOrNull
+import io.hemin.wien.dom.textOrNull
+import io.hemin.wien.dom.toHrefOnlyImageBuilder
+import io.hemin.wien.dom.toITunesCategoryBuilder
 import io.hemin.wien.parser.NamespaceParser
 import io.hemin.wien.util.FeedNamespace
-import io.hemin.wien.util.getAttributeValueByName
 import org.w3c.dom.Node
 
 /**
