+++ src/main/kotlin/io/hemin/wien/parser/namespace/PodloveSimpleChapterParser.kt
@@ -5,8 +5,8 @@ import io.hemin.wien.builder.episode.EpisodePodloveSimpleChapterBuilder
 import io.hemin.wien.builder.podcast.PodcastBuilder
 import io.hemin.wien.parser.NamespaceParser
 import io.hemin.wien.util.FeedNamespace
-import io.hemin.wien.util.asListOfNodes
-import io.hemin.wien.util.getAttributeValueByName
+import io.hemin.wien.dom.asListOfNodes
+import io.hemin.wien.dom.getAttributeValueByName
 import org.w3c.dom.Node
 
 /**
