+++ src/main/kotlin/io/hemin/wien/writer/namespace/BitloveWriter.kt
@@ -1,10 +1,10 @@
 package io.hemin.wien.writer.namespace
 
+import io.hemin.wien.dom.findElementByName
+import io.hemin.wien.dom.setAttributeWithNS
 import io.hemin.wien.model.Episode
 import io.hemin.wien.model.Podcast
 import io.hemin.wien.util.FeedNamespace
-import io.hemin.wien.util.findElementByName
-import io.hemin.wien.util.setAttributeWithNS
 import io.hemin.wien.writer.NamespaceWriter
 import org.w3c.dom.Element
