+++ src/main/kotlin/io/hemin/wien/writer/namespace/GooglePlayWriter.kt
@@ -1,13 +1,13 @@
 package io.hemin.wien.writer.namespace
 
+import io.hemin.wien.dom.appendElement
+import io.hemin.wien.dom.appendHrefOnlyImageElement
+import io.hemin.wien.dom.appendITunesCategoryElements
+import io.hemin.wien.dom.appendYesElementIfTrue
 import io.hemin.wien.model.Episode
 import io.hemin.wien.model.GooglePlayBase
 import io.hemin.wien.model.Podcast
 import io.hemin.wien.util.FeedNamespace
-import io.hemin.wien.util.appendITunesCategoryElements
-import io.hemin.wien.util.appendElement
-import io.hemin.wien.util.appendHrefOnlyImageElement
-import io.hemin.wien.util.appendYesElementIfTrue
 import io.hemin.wien.writer.NamespaceWriter
 import org.w3c.dom.Element
