+++ src/main/kotlin/io/hemin/wien/writer/namespace/ITunesWriter.kt
@@ -1,15 +1,15 @@
 package io.hemin.wien.writer.namespace
 
+import io.hemin.wien.dom.appendElement
+import io.hemin.wien.dom.appendHrefOnlyImageElement
+import io.hemin.wien.dom.appendITunesCategoryElements
+import io.hemin.wien.dom.appendPersonElement
+import io.hemin.wien.dom.appendTrueFalseElement
+import io.hemin.wien.dom.appendYesElementIfTrue
 import io.hemin.wien.model.Episode
 import io.hemin.wien.model.ITunesBase
 import io.hemin.wien.model.Podcast
 import io.hemin.wien.util.FeedNamespace
-import io.hemin.wien.util.appendITunesCategoryElements
-import io.hemin.wien.util.appendElement
-import io.hemin.wien.util.appendHrefOnlyImageElement
-import io.hemin.wien.util.appendPersonElement
-import io.hemin.wien.util.appendTrueFalseElement
-import io.hemin.wien.util.appendYesElementIfTrue
 import io.hemin.wien.writer.NamespaceWriter
 import org.w3c.dom.Element
