+++ src/main/kotlin/io/hemin/wien/writer/namespace/RssWriter.kt
@@ -1,12 +1,12 @@
 package io.hemin.wien.writer.namespace
 
+import io.hemin.wien.dom.appendElement
+import io.hemin.wien.dom.appendRssCategoryElements
+import io.hemin.wien.dom.appendRssImageElement
 import io.hemin.wien.model.Episode
 import io.hemin.wien.model.Podcast
 import io.hemin.wien.util.BooleanStringStyle
 import io.hemin.wien.util.FeedNamespace
-import io.hemin.wien.util.appendRssCategoryElements
-import io.hemin.wien.util.appendElement
-import io.hemin.wien.util.appendRssImageElement
 import io.hemin.wien.util.asBooleanString
 import io.hemin.wien.writer.NamespaceWriter
 import org.w3c.dom.Element
