+++ src/test/kotlin/io/hemin/wien/builder/fake/episode/FakeEpisodeITunesBuilder.kt
@@ -1,7 +1,6 @@
 package io.hemin.wien.builder.fake.episode
 
 import io.hemin.wien.builder.HrefOnlyImageBuilder
-import io.hemin.wien.builder.RssImageBuilder
 import io.hemin.wien.builder.episode.EpisodeITunesBuilder
 import io.hemin.wien.builder.fake.FakeBuilder
 import io.hemin.wien.model.Episode
