+++ src/test/kotlin/io/hemin/wien/builder/validating/episode/ValidatingEpisodeITunesBuilderTest.kt
@@ -8,7 +8,6 @@ import assertk.assertions.isNotNull
 import assertk.assertions.isNull
 import assertk.assertions.prop
 import io.hemin.wien.builder.validating.ValidatingHrefOnlyImageBuilder
-import io.hemin.wien.builder.validating.ValidatingRssImageBuilder
 import io.hemin.wien.model.Episode
 import org.junit.jupiter.api.Test
