+++ src/test/kotlin/io/hemin/wien/parser/NamespaceParserTest.kt
@@ -2,7 +2,7 @@ package io.hemin.wien.parser
 
 import io.hemin.wien.builder.fake.episode.FakeEpisodeBuilder
 import io.hemin.wien.builder.fake.podcast.FakePodcastBuilder
-import io.hemin.wien.util.asListOfNodes
+import io.hemin.wien.dom.asListOfNodes
 import org.w3c.dom.Node
 
 internal abstract class NamespaceParserTest {
